Invert the extras configuration for poppler
libpoppler.so is the internal library that has no stable API/ABI. Only
libpoppler-cpp, libpoppler-glib and libpoppler-qt5 have API
guarantees. So make those the primary API and leave the headers for
the internal one in the -extras package.

We have a single package that uses the internal API in Clear
(cups-filter) and that is part of the desktop bundle.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>